Split TraceabilityMandGs so it hold convenience functions/predefined filters, and TraceabilityMatrix which holds infrastructure.
